undid file seperation, first compilable version
