--- Console.cpp
@@ -1,5 +1,0 @@
-#include "Console.h"
-
-void PrintToConsole(std::string message) {
-
-}
