--- Console.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include<string>
-
-void PrintToConsole(std::string message);
